Extract logic from cmd package
