+++ cmd/payout.go
@@ -1,15 +1,11 @@
 package cmd
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/NodeFactoryIo/vedran/internal/controllers"
-	"github.com/NodeFactoryIo/vedran/internal/models"
-	"github.com/NodeFactoryIo/vedran/internal/payout"
+	"github.com/NodeFactoryIo/vedran/internal/script"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"net/http"
 	"strconv"
 	"strings"
 )
@@ -65,51 +61,11 @@ func init() {
 func payoutCommand(_ *cobra.Command, _ []string) {
 	DisplayBanner()
 	fmt.Println("Payout script running...")
-
-	stats, err := fetchStatsFromEndpoint(loadbalancerUrl + "/api/v1/stats")
+	err := script.ExecutePayout(secret, totalRewardAsFloat64, loadbalancerUrl)
 	if err != nil {
-		log.Errorf("Unable to fetch stats from loadbalancer, because of %v", err)
+		log.Errorf("Unable to execute payout, because of %v", err)
 		return
 	}
-
-	// calculate distribution
-	nodeStatsDetails := make(map[string]models.NodeStatsDetails, len(stats.Stats))
-	for nodeId, nodeStats := range stats.Stats {
-		nodeStatsDetails[nodeId] = nodeStats.Stats
-	}
-
-	distributionByNode := payout.CalculatePayoutDistributionByNode(
-		nodeStatsDetails,
-		totalRewardAsFloat64,
-		float64(stats.Fee),
-	)
-
-	transactionDetails, err := payout.ExecuteAllPayoutTransactions(
-		distributionByNode,
-		stats.Stats,
-		secret,
-		loadbalancerUrl,
-	)
-	if err != nil {
-		log.Errorf("Unable to execute payout transactions, because of %v", err)
-		return
-	}
-
-	// todo - prettify displaying transactions status
-	log.Info(transactionDetails)
+	log.Info("Payout execution finished")
 }
 
-func fetchStatsFromEndpoint(endpoint string) (*controllers.StatsResponse, error) {
-	resp, err := http.Get(endpoint)
-	if err != nil {
-		return nil, err
-	}
-	dec := json.NewDecoder(resp.Body)
-	dec.DisallowUnknownFields()
-	stats := controllers.StatsResponse{}
-	err = dec.Decode(&stats)
-	if err != nil {
-		return nil, err
-	}
-	return &stats, nil
-}
